Refactor email sending code to reusable location.
