+++ api/resources_portal/views/email_invitation.py
@@ -1,20 +1,14 @@
 import json
-import os
-from email.mime.application import MIMEApplication
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
 from pathlib import Path
 
 from django.conf import settings
-from django.http import HttpResponse, HttpResponseForbidden
 from rest_framework import status
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-import boto3
-
 from resources_portal.config.logging import get_and_configure_logger
+from resources_portal.emailer import send_mail
 
 logger = get_and_configure_logger(__name__)
 
@@ -24,75 +18,6 @@ EMAIL_HTML_BODY = (
     .read_text()
     .replace("\n", "")
 )
-
-
-def create_multipart_message(
-    sender: str,
-    recipients: list,
-    title: str,
-    text: str = None,
-    html: str = None,
-    attachments: list = None,
-) -> MIMEMultipart:
-    """
-    Creates a MIME multipart message object.
-    Taken from: https://stackoverflow.com/a/52105406/6095378
-    Uses only the Python `email` standard library.
-    Emails, both sender and recipients, can be just the email string or have the format 'The Name <the_email@host.com>'.
-
-    :param sender: The sender.
-    :param recipients: List of recipients. Needs to be a list, even if only one recipient.
-    :param title: The title of the email.
-    :param text: The text version of the email body (optional).
-    :param html: The html version of the email body (optional).
-    :param attachments: List of files to attach in the email.
-    :return: A `MIMEMultipart` to be used to send the email.
-    """
-    multipart_content_subtype = "alternative" if text and html else "mixed"
-    msg = MIMEMultipart(multipart_content_subtype)
-    msg["Subject"] = title
-    msg["From"] = sender
-    msg["To"] = ", ".join(recipients)
-
-    # Record the MIME types of both parts - text/plain and text/html.
-    # According to RFC 2046, the last part of a multipart message, in this case the HTML message, is best and preferred.
-    if text:
-        part = MIMEText(text, "plain")
-        msg.attach(part)
-    if html:
-        part = MIMEText(html, "html")
-        msg.attach(part)
-
-    # Add attachments
-    for attachment in attachments or []:
-        with open(attachment, "rb") as f:
-            part = MIMEApplication(f.read())
-            part.add_header(
-                "Content-Disposition", "attachment", filename=os.path.basename(attachment)
-            )
-            msg.attach(part)
-
-    return msg
-
-
-def send_mail(
-    sender: str,
-    recipients: list,
-    title: str,
-    text: str = None,
-    html: str = None,
-    attachments: list = None,
-) -> dict:
-    """
-    Send email to recipients. Sends one mail to all recipients.
-    Taken from: https://stackoverflow.com/a/52105406/6095378
-    The sender needs to be a verified email in SES.
-    """
-    msg = create_multipart_message(sender, recipients, title, text, html, attachments)
-    ses_client = boto3.client("ses", region_name=settings.AWS_REGION)
-    return ses_client.send_raw_email(
-        Source=sender, Destinations=recipients, RawMessage={"Data": msg.as_string()}
-    )
 
 
 @api_view(["POST"])
@@ -130,4 +55,4 @@ def email_invitation_view(request):
         logger.info(f"In prod an email would have been sent to {email}:")
         print(body)
 
-    return HttpResponse(status=status.HTTP_201_CREATED)
+    return Response(status=status.HTTP_201_CREATED)
